+++ src/os2datascanner/projects/admin/import_services/views/msgraph_config_views.py
@@ -33,6 +33,7 @@ class MSGraphEditForm(forms.ModelForm):
         fields = [
             'organization',
             'grant',
+            'hide_units_on_import'
         ]